Remove extended parental leave from the list
The motivation for giving fathers an extra month of paternity leave in Spain came from the fact that laws in Spain give men 5 weeks and women 16 weeks, which has an effect of women being less competitive in the job market. By paying an extra month to men we aimed to correct this inequality, so for Vizzuality it's not more profitable or convenient to hire men or women, and thus remove the possible bias.

I thought it would be good to clarify the nature of this policy and make sure we are all on the same page about this, also figure out how this policy applies in different places and adapts as laws change.

My proposal, as a company-wide policy is that we try to correct biases when we detect them in each country's laws, not necessarily trying to make them equivalent across countries, but promoting equality within each.

I'd like to use this thread to discuss laws in Spain, UK, and Portugal and decide what our position should be in each country. After a first assessment, laws in Portugal seem to be very gender-balanced (http://www.cgtp.pt/igualdade/perguntas-frequentes/7584-direitos-de-maternidade-e-paternidade-parentalidade-0) so there may not be a need for this policy there. Laws in Spain are gradually getting better (with 8 weeks for men being approved this month and equality expected for next year). I haven't had a chance to study UK's laws.